requirements.txt
h5py
ipykernel
ipython
jupyter
matplotlib==3.6.1
numpy==1.23.4
pickleshare
pre-commit
pyyaml
scikit-learn==1.1.3
scikit-optimize==0.9.0
scipy==1.9.3
ml-collections
wandb
